+++ visualization/graphfilegen/src/main/java/edu/uw/modelab/pojo/Route.java
@@ -2,17 +2,17 @@ package edu.uw.modelab.pojo;
 
 public class Route {
 
-	private final long id;
+	private final int id;
 	private final String name;
 	private final String agencyId;
 
-	public Route(final long id, final String name, final String agencyId) {
+	public Route(final int id, final String name, final String agencyId) {
 		this.id = id;
 		this.name = name;
 		this.agencyId = agencyId;
 	}
 
-	public long getId() {
+	public int getId() {
 		return id;
 	}
