+++ visualization/graphfilegen/src/main/java/edu/uw/modelab/pojo/Stop.java
@@ -2,12 +2,12 @@ package edu.uw.modelab.pojo;
 
 public class Stop {
 
-	private final long id;
+	private final int id;
 	private final String name;
 	private final String lat;
 	private final String lon;
 
-	public Stop(final long id, final String name, final String lat,
+	public Stop(final int id, final String name, final String lat,
 			final String lon) {
 		this.id = id;
 		this.name = name;
@@ -15,7 +15,7 @@ public class Stop {
 		this.lon = lon;
 	}
 
-	public long getId() {
+	public int getId() {
 		return id;
 	}
